[MOV] http_routing,web: move frontend module definitions to web
We have a list of modules whose JavaScript translations should be loaded
on the frontend for the portal, website, etc. The method defining this
list was in the `http_routing` module. However, not all the modules
loaded in the frontend depend on `http_router` (like `html_editor`).

In this commit we move the frontend module definitions from the
`http_routing` module to the `web` module, which is an ancestor and
doesn't require any changes in overriding modules.

Meanwhile we also removed some unused code and moved the controller
route loading the website translations to the `web` module as well.
