+++ addons/http_routing/controllers/main.py
@@ -1,21 +1,8 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import http
-from odoo.http import request
-from odoo.addons.web.controllers.home import Home
+
 from odoo.addons.web.controllers.session import Session
-from odoo.addons.web.controllers.webclient import WebClient
-
-
-class Routing(Home):
-
-    @http.route('/website/translations', type='http', auth="public", readonly=True, sitemap=False)
-    def get_website_translations(self, hash=None, lang=None, mods=None):
-        IrHttp = request.env['ir.http'].sudo()
-        modules = IrHttp.get_translation_frontend_modules()
-        if mods:
-            modules += mods.split(',')
-        return WebClient().translations(hash, mods=','.join(modules), lang=lang)
 
 
 class SessionWebsite(Session):
